% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getp.R
\name{getp.func}
\alias{getp.func}
\title{This function caculate P-value for every trio}
\usage{
getp.func(
  i,
  triomatrix,
  confounders,
  Minperm = 100,
  Maxperm = 10000,
  use.gpd = FALSE,
  gpd.perm = 0.01,
  pool_cov = NULL,
  est_conf_pool_idx = NULL,
  use.PC = FALSE
)
}
\arguments{
\item{i}{Trios index in triomatrix}

\item{triomatrix}{A three-dimensional matrix of size: samples number * trios
number * 3. Triomatrix[i,j,1] represents the genotype of the j-th trios at
the i-th sample,and triomatrix[i,j,2] represents the feature1 data of the
j-th trios at the i-th sample, triomatrix[i,j,3] represents the feature2
data of the j-th trios at the i-th sample.}

\item{Minperm}{Decide whether to use the parameters of the GPD fit. If the
value is 0, only the nominal P-value is calculated. If the proportion of
the permutation statistic better than the original statistic to the total
number of permutations exceeds this value, a more accurate empirical P
value is estimated using the GPD fit. If \code{Minperm}=0, only the nominal
P-value is calculated. We set \code{Minperm}=100 as default.}

\item{Maxperm}{Maximum number of permutation. We set \code{Maxperm}=10000 as
default.}

\item{use.gpd}{Whether to use the GPD fit to estimate a more accurate
empirical P-value. We set \code{use.gpd}=NULL as default.}

\item{gpd.perm}{The proportion parameter for estimating the empirical P-value
when using GPD fit. When the proportion of permutation better than the
original statistic is greater than par, the GPD is fitted to estimate the
empirical P-value. We set \code{gpd.perm}=0.01 as default.}

\item{pool_cov}{Candidate Confusion Variable Pool. We set
\code{pool_cov}=NULL as default, which \code{use.PC} requied.}

\item{est_conf_pool_idx}{The index of the adaptively selected confunding
variable. We set \code{pool_cov}=NULL as default, at this time
\code{pool_cov}=NULL too.}

\item{use.PC}{Whether the candidate confusion variable pool is PCs.}

\item{confunders}{A confounders matrix which is adjusted in all mediation
tests.}
}
\value{
The algorithm will return a list of nperm, empirical.p,
  empirical.p.gpd, nominal.p, std.error, t_stat, beta, beta.total,
  beta.change. \item{nperm}{The actual number of permutations for testing
  mediation, equal to the input parameter nperm.} \item{empirical.p}{The
  mediation Empirical P-values with nperm times permutation.}
  \item{empirical.p.gpd}{The mediation Empirical P-values with nperm times
  permutation using GPD fit.} \item{nominal.p}{The mediation nominal
  P-values. A matrix with dimension of the number of trios.}
  \item{std.error}{The return std.error value of feature1 for fit liner
  models.} \item{t_stat}{The return t_stat value of feature1 for fit liner
  models.} \item{beta}{The return beta value of feature2 for fit liner models
  in the case of feature1.} \item{beta.total}{The return beta value of
  feature2 for fit liner models without considering feature1.}
  \item{beta.change}{The proportions mediated.}
}
\description{
This function calculates the P-value for each trios. If
  \code{Minperm}=0, only the nominal P-value is calculated. If
  \code{Minperm}=\code{Maxperm}, the empirical P-value is calculated using a
  fixed number of permutation statistics; otherwise, the empirical P-value is
  calculated using the adaptive permutation scheme. The user can specify
  whether to use the GPD fit to estimate a more accurate empirical P-value,
  and at the same time specify how small the empirical P-value is for GPD
  fitting.
}
