% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmfp.R
\name{gmfp}
\alias{gmfp}
\title{Genomic Mediation analysis with Fixed Permutation scheme}
\usage{
gmfp(snp.dat, fea.dat, conf, trios.idx, cl = NULL, nperm = 10000)
}
\arguments{
\item{snp.dat}{The eQTL genotype matrix. Each row is an eQTL, each column is
a sample.}

\item{fea.dat}{A feature profile matrix. Each row is for one feature, each
column is a sample.}

\item{conf}{A confounders matrix which is adjusted in mediation tests. Each
row is a confounder, each column is a sample.}

\item{trios.idx}{A matrix of selected trios indexes (row numbers) for
mediation tests. Each row consists of the index (i.e., row number) of the
eQTL in \code{snp.dat}, the index of cis-gene feature in \code{fea.dat},
and the index of trans-gene feature in \code{fea.dat}. The dimension is the
number of trios by three.}

\item{cl}{Parallel backend if it is set up. It is used for parallel
computing. We set \code{cl}=NULL as default.}

\item{nperm}{The number of permutations for testing mediation. If
\code{nperm}=0, only the nominal P-value is calculated. We set
\code{nperm}=10000 as default.}
}
\value{
The algorithm will return a list of empirical.p, nominal.p, beta,
  std.error, t_stat, beta.total, beta.change. \item{empirical.p}{The
  mediation empirical P-values with nperm times permutation. A matrix with
  dimension of the number of trios.} \item{nominal.p}{The mediation nominal
  P-values. A matrix with dimension of the number of trios.}
  \item{std.error}{The return std.error value of feature1 for fit liner
  models. A matrix with dimension of the number of trios.} \item{t_stat}{The
  return t_stat value of feature1 for fit liner models. A matrix with
  dimension of the number of trios.} \item{beta}{The return beta value of
  feature2 for fit liner models in the case of feature1. A matrix with
  dimension of the number of trios.} \item{beta.total}{The return beta value
  of feature2 for fit liner models without considering feature1. A matrix
  with dimension of the number of trios.} \item{beta.change}{The proportions
  mediated. A matrix with dimension of the number of trios.}
}
\description{
The gmfp function performs genomic mediation analysis with fixed
  permutation. It tests for mediation effects for a set of user specified
  mediation trios(e.g., eQTL, cis- and trans-genes) in the genome with the
  assumption of the presence of cis-association.

  It returns the mediation p-values(nominal and empirical), the coefficient
  of linear models(e.g, t_stat, std.error, beta, beta.total) and the
  proportions mediated(e.g., the percentage of reduction in trans-effects
  after accounting for cis-mediation).
}
\details{
The function performs genomic mediation analysis with fixed
  permutation. \code{Fixed Permutation scheme}{When calculating the empirical
  P-value, the data is permutated by a fixed number of times, and the
  statistics after permutation are separately calculated. Assuming that the
  number of permutation is N, where the number of permutation statistics that
  is better than the original statistic is M, then the Empirical P-value = (M
  + 1) / (N + 1).}
}
\examples{

output <- gmfp(conf = dat$known.conf, fea.dat = dat$fea.dat, snp.dat = dat$snp.dat,
               trios.idx = dat$trios.idx[1:10,], nperm = 100)

\dontrun{
  ## generate a cluster with 2 nodes for parallel computing
  cl <- makeCluster(2)
  output <- gmfp(conf = dat$known.conf, fea.dat = dat$fea.dat, snp.dat = dat$snp.dat,
                 trios.idx = dat$trios.idx[1:10,], cl = cl, nperm = 100)
  stopCluster(cl)
}

}
