% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpdfit.R
\name{gpdfit}
\alias{gpdfit}
\title{Fitting the generalized pareto distribution using 'ML'}
\usage{
gpdfit(z)
}
\value{
Pharmat  estimated shape and scale parameter
}
\description{
Fitting the generalized pareto distribution using 'ML'
}
