% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppermest.R
\name{Pgpd}
\alias{Pgpd}
\title{Computing permutation test P-value of the GPD approximation using 'ML'}
\usage{
Pgpd(y, x0, N, Nexc)
}
\value{
Phat  estimated P-value
}
\description{
Computing permutation test P-value of the GPD approximation using 'ML'
}
