% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ppermest.R
\name{Ppermest}
\alias{Ppermest}
\title{Estimation of P-value using Generalized Pareto Distribution(GPD)}
\usage{
Ppermest(x0, y, Nexcmax = 250, proportion = 0.01)
}
\value{
P  Estimate of the P-value.
}
\description{
Estimation of P-value using Generalized Pareto Distribution(GPD)
}
