% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvalue.R
\name{qvalue}
\alias{qvalue}
\title{qvalue is the function to estimate q values given a vector of p values.}
\usage{
qvalue(p, fdr.level = NULL, pfdr = FALSE, lfdr.out = TRUE, pi0 = NULL, ...)
}
\description{
The code for this function is extracted from the "qvalue" package written by
Alan Dabney and John Storey. We include the functions of qvalue here since it
is reported than some operating system cannot isntall it, and we are using
version 2.8.0.
}
